move logic outsite management command
